refactor(CMS): move salon files into component dirs, remove dead code
- Move salon/cell, salon/changelogs, salon/communities, salon/posts into respective component dirs
- Delete unused Docs/Tree component and its salon/docs/tree
- Delete unused Docs/Table.tsx, Docs/index.tsx, salon/docs/*
- Remove now-empty central CMS/salon directory
