--- frontend/core/unit/DashboardThread/CMS/Docs/index.tsx
@@ -1,3 +1,0 @@
-export default function Docs() {
-  return null
-}
