+++ frontend/core/unit/DashboardThread/CMS/Posts/salon/index.ts
